refactor agent and openai client config
